move launchPrep route
